Attempt to use generic to benchmark 2020 day 2 puzzles
This has a compiler error that has something to do
with the fact that the `Intermediate` type argument
contains a reference with a specific lifetime.

I haven't been able to figure out how to tell the compiler
exactly how to resolve the types and lifetimes to make
it all fit together.